+++ CHANGELOG.md
@@ -70,6 +70,7 @@ Users may have to change how they access the system
 
 ### Corrected Bugs
 
+* Fixed: Performance enhancement of unique list check (of duplicate keys)
 * Fixed: Validate/commit error with false positive yang choice changes detected in validation found in ietf-ipfix-psamp.yang.
 * Fixed: Accepted added subtrees containing lists with duplicate keys.
 * Fixed: [default state data returned with get-config](https://github.com/clicon/clixon/issues/140)